add global mock db
