+++ dm/master/server_test.go
@@ -16,7 +16,6 @@ package master
 import (
 	"bytes"
 	"context"
-	"database/sql"
 	"fmt"
 	"io/ioutil"
 	"net"
@@ -441,22 +440,6 @@ func generateServerConfig(c *check.C, name string) *Config {
 	return cfg1
 }
 
-// db use for remove data
-// verDB user for show version.
-type mockDBProvider struct {
-	verDB *sql.DB
-	db    *sql.DB
-}
-
-// return db if verDB was closed.
-func (d *mockDBProvider) Apply(config config.DBConfig) (*conn.BaseDB, error) {
-	if err := d.verDB.Ping(); err != nil {
-		// nolint:nilerr
-		return conn.NewBaseDB(d.db, func() {}), nil
-	}
-	return conn.NewBaseDB(d.verDB, func() {}), nil
-}
-
 func (t *testMaster) TestQueryStatus(c *check.C) {
 	ctrl := gomock.NewController(c)
 	defer ctrl.Finish()
@@ -719,7 +702,7 @@ func (t *testMaster) TestCheckTask(c *check.C) {
 	var wg sync.WaitGroup
 	ctx, cancel := context.WithCancel(context.Background())
 	server.scheduler, _ = t.testMockScheduler(ctx, &wg, c, sources, workers, "", t.workerClients)
-	mock := t.initVersionDB(c)
+	mock := conn.InitVersionDB(c)
 	defer func() {
 		conn.DefaultDBProvider = &conn.DefaultDBProviderImpl{}
 	}()
@@ -742,7 +725,7 @@ func (t *testMaster) TestCheckTask(c *check.C) {
 	// simulate invalid password returned from scheduler, but config was supported plaintext mysql password, so cfg.SubTaskConfigs will success
 	ctx, cancel = context.WithCancel(context.Background())
 	server.scheduler, _ = t.testMockScheduler(ctx, &wg, c, sources, workers, "invalid-encrypt-password", t.workerClients)
-	mock = t.initVersionDB(c)
+	mock = conn.InitVersionDB(c)
 	mock.ExpectQuery("SHOW GLOBAL VARIABLES LIKE 'version'").WillReturnRows(sqlmock.NewRows([]string{"Variable_name", "Value"}).
 		AddRow("version", "5.7.25-TiDB-v4.0.2"))
 	resp, err = server.CheckTask(context.Background(), &pb.CheckTaskRequest{
@@ -778,7 +761,7 @@ func (t *testMaster) TestStartTask(c *check.C) {
 	}
 	server.scheduler, _ = t.testMockScheduler(ctx, &wg, c, sources, workers, "",
 		makeWorkerClientsForHandle(ctrl, taskName, sources, workers, req))
-	mock := t.initVersionDB(c)
+	mock := conn.InitVersionDB(c)
 	defer func() {
 		conn.DefaultDBProvider = &conn.DefaultDBProviderImpl{}
 	}()
@@ -798,7 +781,7 @@ func (t *testMaster) TestStartTask(c *check.C) {
 
 	// check start-task with an invalid source
 	invalidSource := "invalid-source"
-	mock = t.initVersionDB(c)
+	mock = conn.InitVersionDB(c)
 	mock.ExpectQuery("SHOW GLOBAL VARIABLES LIKE 'version'").WillReturnRows(sqlmock.NewRows([]string{"Variable_name", "Value"}).
 		AddRow("version", "5.7.25-TiDB-v4.0.2"))
 	resp, err = server.StartTask(context.Background(), &pb.StartTaskRequest{
@@ -819,7 +802,7 @@ func (t *testMaster) TestStartTask(c *check.C) {
 	defer func() {
 		checker.CheckSyncConfigFunc = bakCheckSyncConfigFunc
 	}()
-	mock = t.initVersionDB(c)
+	mock = conn.InitVersionDB(c)
 	mock.ExpectQuery("SHOW GLOBAL VARIABLES LIKE 'version'").WillReturnRows(sqlmock.NewRows([]string{"Variable_name", "Value"}).
 		AddRow("version", "5.7.25-TiDB-v4.0.2"))
 	resp, err = server.StartTask(context.Background(), &pb.StartTaskRequest{
@@ -830,28 +813,6 @@ func (t *testMaster) TestStartTask(c *check.C) {
 	c.Assert(resp.Result, check.IsFalse)
 	c.Assert(resp.Msg, check.Matches, errCheckSyncConfigReg)
 	t.clearSchedulerEnv(c, cancel, &wg)
-}
-
-func (t *testMaster) initVersionDB(c *check.C) sqlmock.Sqlmock {
-	db, mock, err := sqlmock.New()
-	c.Assert(err, check.IsNil)
-	if mdbp, ok := conn.DefaultDBProvider.(*mockDBProvider); ok {
-		mdbp.verDB = db
-	} else {
-		conn.DefaultDBProvider = &mockDBProvider{verDB: db}
-	}
-	return mock
-}
-
-func (t *testMaster) initMockDB(c *check.C) sqlmock.Sqlmock {
-	db, mock, err := sqlmock.New()
-	c.Assert(err, check.IsNil)
-	if mdbp, ok := conn.DefaultDBProvider.(*mockDBProvider); ok {
-		mdbp.db = db
-	} else {
-		conn.DefaultDBProvider = &mockDBProvider{db: db}
-	}
-	return mock
 }
 
 func (t *testMaster) TestStartTaskWithRemoveMeta(c *check.C) {
@@ -900,13 +861,13 @@ func (t *testMaster) TestStartTaskWithRemoveMeta(c *check.C) {
 	c.Assert(server.pessimist.Start(ctx, t.etcdTestCli), check.IsNil)
 	c.Assert(server.optimist.Start(ctx, t.etcdTestCli), check.IsNil)
 
-	verMock := t.initVersionDB(c)
+	verMock := conn.InitVersionDB(c)
 	defer func() {
 		conn.DefaultDBProvider = &conn.DefaultDBProviderImpl{}
 	}()
 	verMock.ExpectQuery("SHOW GLOBAL VARIABLES LIKE 'version'").WillReturnRows(sqlmock.NewRows([]string{"Variable_name", "Value"}).
 		AddRow("version", "5.7.25-TiDB-v4.0.2"))
-	mock := t.initMockDB(c)
+	mock := conn.InitMockDB(c)
 	mock.ExpectBegin()
 	mock.ExpectExec(fmt.Sprintf("DROP TABLE IF EXISTS `%s`.`%s`", cfg.MetaSchema, cputil.LoaderCheckpoint(cfg.Name))).WillReturnResult(sqlmock.NewResult(1, 1))
 	mock.ExpectExec(fmt.Sprintf("DROP TABLE IF EXISTS `%s`.`%s`", cfg.MetaSchema, cputil.SyncerCheckpoint(cfg.Name))).WillReturnResult(sqlmock.NewResult(1, 1))
@@ -921,7 +882,7 @@ func (t *testMaster) TestStartTaskWithRemoveMeta(c *check.C) {
 		defer wg.Done()
 		time.Sleep(10 * time.Microsecond)
 		// start another same task at the same time, should get err
-		verMock2 := t.initVersionDB(c)
+		verMock2 := conn.InitVersionDB(c)
 		verMock2.ExpectQuery("SHOW GLOBAL VARIABLES LIKE 'version'").WillReturnRows(sqlmock.NewRows([]string{"Variable_name", "Value"}).
 			AddRow("version", "5.7.25-TiDB-v4.0.2"))
 		resp1, err1 := server.StartTask(context.Background(), req)
@@ -995,10 +956,10 @@ func (t *testMaster) TestStartTaskWithRemoveMeta(c *check.C) {
 	err = server.optimist.Start(ctx, t.etcdTestCli)
 	c.Assert(err, check.IsNil)
 
-	verMock = t.initVersionDB(c)
+	verMock = conn.InitVersionDB(c)
 	verMock.ExpectQuery("SHOW GLOBAL VARIABLES LIKE 'version'").WillReturnRows(sqlmock.NewRows([]string{"Variable_name", "Value"}).
 		AddRow("version", "5.7.25-TiDB-v4.0.2"))
-	mock = t.initMockDB(c)
+	mock = conn.InitMockDB(c)
 	mock.ExpectBegin()
 	mock.ExpectExec(fmt.Sprintf("DROP TABLE IF EXISTS `%s`.`%s`", cfg.MetaSchema, cputil.LoaderCheckpoint(cfg.Name))).WillReturnResult(sqlmock.NewResult(1, 1))
 	mock.ExpectExec(fmt.Sprintf("DROP TABLE IF EXISTS `%s`.`%s`", cfg.MetaSchema, cputil.SyncerCheckpoint(cfg.Name))).WillReturnResult(sqlmock.NewResult(1, 1))
@@ -1013,7 +974,7 @@ func (t *testMaster) TestStartTaskWithRemoveMeta(c *check.C) {
 		defer wg.Done()
 		time.Sleep(10 * time.Microsecond)
 		// start another same task at the same time, should get err
-		vermock2 := t.initVersionDB(c)
+		vermock2 := conn.InitVersionDB(c)
 		vermock2.ExpectQuery("SHOW GLOBAL VARIABLES LIKE 'version'").WillReturnRows(sqlmock.NewRows([]string{"Variable_name", "Value"}).
 			AddRow("version", "5.7.25-TiDB-v4.0.2"))
 		resp1, err1 := server.StartTask(context.Background(), req)
@@ -1098,7 +1059,7 @@ func (t *testMaster) TestOperateTask(c *check.C) {
 	sourceResps := []*pb.CommonWorkerResponse{{Result: true, Source: sources[0]}, {Result: true, Source: sources[1]}}
 	server.scheduler, _ = t.testMockScheduler(ctx, &wg, c, sources, workers, "",
 		makeWorkerClientsForHandle(ctrl, taskName, sources, workers, startReq, pauseReq, resumeReq, stopReq1, stopReq2))
-	mock := t.initVersionDB(c)
+	mock := conn.InitVersionDB(c)
 	defer func() {
 		conn.DefaultDBProvider = &conn.DefaultDBProviderImpl{}
 	}()
@@ -1941,7 +1902,7 @@ func (t *testMaster) TestGetCfg(c *check.C) {
 	server.etcdClient = t.etcdTestCli
 
 	// start task
-	mock := t.initVersionDB(c)
+	mock := conn.InitVersionDB(c)
 	defer func() {
 		conn.DefaultDBProvider = &conn.DefaultDBProviderImpl{}
 	}()
